+++ CHANGELOG.md
@@ -19,8 +19,10 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #985 Added function to filter notifications for contractAgreementIds
 - #786 Added authorization as admin for submodel api & registry api
 - #884 Upgraded tractionBatteryCode from 1.0.0 to 2.0.0
+- #1009 reimplemented retry request logic for notification approval
 - #786 Added alternative port (only accessible within same cluster) for application which is used for unsecured API endpoints.
 - #994 improved bpn edc configuration view uux
+
 
 ### Added
 - #832 added policymanagement list view, creator and editor